Oberoi Kolshet Road Thane
Oberoi Kolshet Road is a new launch housing property that is an address to illustrate luxury. This luxurious development is bought by Oberoi Realty. It is wrapped in magnificence and amidst nature that will gift you a passive and happy life.

This residency is a perfect escape from the busy life of the city and welcomes with open arms a pristine gated community. This stylishly designed venture is a distinctive and different development with a high-end living contribution. This private space is befitting royalty with splendidly designed homes. Provident Wagholi provides you with 2, 3 and 4 BHK Luxury Apartments.

This stillness and calm houses are a place to live in peace with loved ones. The flats are intended in a way to have the best pleasure of space with less loading and more space.

For More Details :
Visit - https://oberoi.developerprojects.in/thane/oberoi-kolshet-road/